Enum cleanup.
- Moved PaginationDeletion into its own file.
- Moved SplitType into its own file under the Enum namespace.
- Removed unnecessary Flags attribute on PollBehaviour.
- Improved enum XML documentation.
