+++ DSharpPlus.Interactivity/Enums/SplitType.cs
@@ -1,0 +1,18 @@
+﻿namespace DSharpPlus.Interactivity.Enums
+{
+    /// <summary>
+    /// Specifies how to split a string.
+    /// </summary>
+    public enum SplitType
+    {
+        /// <summary>
+        /// Splits string per 500 characters.
+        /// </summary>
+        Character,
+
+        /// <summary>
+        /// Splits string per 15 lines.
+        /// </summary>
+        Line
+    }
+}
